requirements.txt
openai>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
prophet>=1.1.5
textblob>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.15.0
wordcloud>=1.9.2
google-generativeai>=0.3.0